Implement _switch_branch function
With this function it is now possible to call the script with the
manifest branch option -b (same as used in the repo script). Before
doing so, ensure that the target_branch is already available for the
.repo/local_manfests repository, in case the target branch is not
available the switch will fail.

Providing the -b will not only switch the repository but will also build
it the same way as it is done without providing the -b switch.